refactor => separate handle clients
